chore(typing): bring six more subpackages to mypy --strict
Extends the strict-typing effort started on molforge.core. Six more
subpackages — io, sequence, structure, metrics, ensembles,
validation — now pass mypy --strict with zero errors, and the CI
gate is widened to enforce all seven.

Context: where this fits

The pre-1.0 API audit produced a documented public surface; the
typing work makes that surface trustworthy. molforge.core went
strict-clean. The full tree still showed 48 strict errors
across 19 files, surfaced by the non-blocking informational
`mypy src` CI step. This commit clears the next tranche.

Why these six and not the biggest pile

The 12 errors
Most were benign and shared two patterns:
  - no-any-return (tm.py, geometry.py x2, contacts.py, weighting.py,
    gdt.py): numpy operations — .sum/.mean/.astype/linalg.norm and
    a float-exponentiation — that mypy widens to Any, against a
    function declared to return a concrete ndarray or float. Fixed
    with explicit casts (and one float() wrap) that document the
    dtype we already know the operation produces. No behaviour
    change.
  - unused type: ignore (orchestration.py, dispatch.py): two stale
    ignores that newer mypy correctly flags as no longer needed.
    Removed.
  - var-annotated (dockq.py): an empty list `sel = []` mypy can't
    infer; annotated `sel: list[int] = []`.

Two were genuine annotation bugs, not noise:
  - structure/dssp.py: _place_hydrogens was declared
    `-> NDArray[np.float32]` but actually returns a
    `(h_coords, h_mask)` tuple — and its sole caller unpacks two
    values from it. The annotation simply lied. Corrected to
    `tuple[NDArray[np.float32], NDArray[np.bool_]]`.
  - sequence/alignment.py: _score was declared `-> NDArray[np.int_]`
    but builds its array with `dtype=np.int32`. np.int_ is
    platform-dependent (int64 on 64-bit Linux/macOS), so the
    declared and actual dtypes genuinely disagreed. Corrected to
    NDArray[np.int32] to match what the function builds.

  weighting.py also had an attr-defined error: _coerce_to_score_array
  does `first = seq[0]` (seq is list[object]) then accesses
  `.score` after a hasattr guard. mypy can't narrow object on
  hasattr; switched the access to getattr, which is type-clean and
  preserves the exact runtime behaviour.

Enforcement widened
  - CI: the strict gate step (previously `mypy --strict
    src/molforge/core/`) now lists all seven clean subpackages. The
    non-blocking informational `mypy src` step is unchanged and
    continues to track the not-yet-strict subpackages (ml, plugins,
    wrappers).

  - Regression test: tests/unit/core/test_typing.py is moved up to
    tests/unit/test_typing.py (it no longer belongs under core/) and
    parametrized over all seven subpackages. Still slow-marked,
    still skips cleanly without mypy installed. The subpackage list
    in the test has a comment noting it must stay in sync with the
    CI gate and the api-stability doc.

  - docs/architecture/api-stability.md: new "Type checking" section
    listing the seven strict-clean subpackages and noting ml /
    plugins / wrappers as in-progress.

All 486 tests across the six affected subpackages pass unchanged —
the fixes are annotations and casts, no behaviour change. Test count
924 passed + 11 skipped (the typing regression test went from 1 case
to 7 parametrized cases, +6).

+++ src/molforge/metrics/tm.py
@@ -53,7 +53,7 @@ def _d0(length: int) -> float:
     """
     if length < 21:
         return 0.5
-    return 1.24 * (length - 15) ** (1.0 / 3.0) - 1.8
+    return float(1.24 * (length - 15) ** (1.0 / 3.0) - 1.8)
 
 
 def _ca_coords(protein: Protein) -> np.ndarray:
